Add subclass proficiencies and race proficiencies to DataGetter
Unlike class proficiencies, subclass proficiencies are only stored as a string without a cohesive format to parse with. We can parse out most unnecessary information and return a string, from where it will be up to the user to add the proficiencies themselves (we can add an input to do so on the page for choosing class and subclass). Race proficiencies can be parsed for skills, but other types of proficiencies aren't very parse-able, so they've been put into an "other" category so the user can read through and add them, similar to subclass proficiencies.