Add in spellcard source & fix type formatting
Small formatting improvement in type formatting on spellcards by positioning it from the right instead of the left. Since we're not displaying spellcards per class, I'm also using the class div to hold the name of the source material.

There is still an issue where if the text is too long, it doesn't fit on the card and there's nothing to handle shrinking it down or splitting it up. Removed the "example card" element now that spellcard is formatted (outside of this one text shrink/split issue).