Update class & race selection
Keep track of selected class and levels. Fix issue where clicking on the level select dropdown collapses the class text.

Organize images into their own folder, and add images for each race. These images display on the race dropdown. The plan is to add in the race image and some benefits of selecting the race to the right of the image.