Add class icons, race description, display selected race
Add image icons by each of the classes, and create a text description of each race. Update the selected race at the top of the race section with the user's choice. Currently we are missing subclasses and subraces, as well as proper multiclassing logic, so that still needs to be done