Fetch specific-info API route, update stats page
Add some descriptive text and fix the input boxes to only accept valid stat values.

As part of the descriptive text on the stats page, I wanted to include a suggestion for the stats the user should prioritize based on the class they've selected. This information is sent over by the "specific-info" route from our server. As such, I set up the code to fetch the API data when it is needed (when we are entering the section where we transition from general-info to specific-info, if the data isn't already loaded in, we fetch it).